+++ src/euroeval/dataset_configs/__init__.py
@@ -4,6 +4,7 @@ from ..data_models import DatasetConfig
 from ..languages import get_all_languages
 from ..tasks import SPEED
 from ..utils import load_custom_datasets_module
+from .bosnian import *  # noqa: F403
 from .bulgarian import *  # noqa: F403
 from .croatian import *  # noqa: F403
 from .czech import *  # noqa: F403
@@ -25,7 +26,7 @@ from .polish import *  # noqa: F403
 from .portuguese import *  # noqa: F403
 from .serbian import *  # noqa: F403
 from .slovak import *  # noqa: F403
-from .slovenian import *  # noqa: F403
+from .slovene import *  # noqa: F403
 from .spanish import *  # noqa: F403
 from .swedish import *  # noqa: F403
 from .ukrainian import *  # noqa: F403