+++ src/euroeval/prompt_templates/reading_comprehension.py
@@ -4,6 +4,7 @@ import typing as t
 
 from ..data_models import PromptConfig
 from ..languages import (
+    BOSNIAN,
     BULGARIAN,
     CROATIAN,
     CZECH,
@@ -27,7 +28,7 @@ from ..languages import (
     PORTUGUESE,
     SERBIAN,
     SLOVAK,
-    SLOVENIAN,
+    SLOVENE,
     SPANISH,
     SWEDISH,
     UKRAINIAN,
@@ -37,6 +38,18 @@ if t.TYPE_CHECKING:
     from ..languages import Language
 
 RC_TEMPLATES: dict["Language", PromptConfig] = {
+    BOSNIAN: PromptConfig(
+        default_prompt_prefix="Slijede tekstovi s pitanjima i odgovorima.",
+        default_prompt_template=(
+            "Tekst: {text}\nPitanje: {question}\nOdgovor s najviše 3 riječi: {label}"
+        ),
+        default_instruction_prompt=(
+            "Tekst: {text}\n\n"
+            "Odgovorite na sljedeće pitanje o gornjem tekstu s najviše 3 riječi.\n\n"
+            "Pitanje: {question}"
+        ),
+        default_prompt_label_mapping=dict(),
+    ),
     BULGARIAN: PromptConfig(
         default_prompt_prefix="Следват текстове със съответни въпроси и отговори.",
         default_prompt_template="Текст: {text}\nВъпрос: {question}\nОтговор с максимум "
@@ -238,7 +251,7 @@ RC_TEMPLATES: dict["Language", PromptConfig] = {
         "sobre o texto acima num máximo de 3 palavras.\n\nPergunta: {question}",
         default_prompt_label_mapping=dict(),
     ),
-    SLOVENIAN: PromptConfig(
+    SLOVENE: PromptConfig(
         default_prompt_prefix=(
             "Spodaj so besedila z ustreznimi vprašanji in odgovori."
         ),